add some casts
git-svn-id: https://svn.r-project.org/R/trunk@59499 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/tools/src/text.c
@@ -66,8 +66,8 @@ SEXP delim_match(SEXP x, SEXP delims)
 
     delim_start = translateChar(STRING_ELT(delims, 0));
     delim_end = translateChar(STRING_ELT(delims, 1));
-    lstart = strlen(delim_start);
-    lend = strlen(delim_end);
+    lstart = (int)strlen(delim_start);
+    lend = (int)strlen(delim_end);
     equal_start_and_end_delims = strcmp(delim_start, delim_end) == 0;
 
     n = length(x);
@@ -100,7 +100,7 @@ SEXP delim_match(SEXP x, SEXP delims)
                 {
                     if (mbcslocale)
                     {
-                        used = Rf_mbrtowc(NULL, s, MB_CUR_MAX, &mb_st);
+                        used = (int)Rf_mbrtowc(NULL, s, MB_CUR_MAX, &mb_st);
                         if (used == 0)
                             break;
                         s += used;
@@ -134,7 +134,7 @@ SEXP delim_match(SEXP x, SEXP delims)
             }
             if (mbcslocale)
             {
-                used = Rf_mbrtowc(NULL, s, MB_CUR_MAX, &mb_st);
+                used = (int)Rf_mbrtowc(NULL, s, MB_CUR_MAX, &mb_st);
                 if (used == 0)
                     break;
                 s += used;
@@ -276,7 +276,7 @@ SEXP doTabExpand(SEXP strings, SEXP starts) /* does tab expansion for UTF-8 stri
             if (0x80 <= (unsigned char)*input && (unsigned char)*input <= 0xBF)
                 start--;
             else if (*input == '\n')
-                start = buffer - b - 1;
+                start = (int)(buffer - b - 1);
             if (*input == '\t')
                 do
                 {
@@ -286,7 +286,7 @@ SEXP doTabExpand(SEXP strings, SEXP starts) /* does tab expansion for UTF-8 stri
                 *b++ = *input;
             if (b - buffer >= bufsize - 8)
             {
-                int pos = b - buffer;
+                int pos = (int)(b - buffer);
                 char *tmp;
                 bufsize *= 2;
                 tmp = realloc(buffer, bufsize * sizeof(char));
